+++ esmf_regrid/tests/unit/schemes/test_ESMFBilinear.py
@@ -5,6 +5,7 @@ import pytest
 from esmf_regrid.schemes import ESMFBilinear
 from esmf_regrid.tests.unit.schemes.__init__ import (
     _test_cube_regrid,
+    _test_dtype_handling,
     _test_invalid_mdtol,
     _test_mask_from_init,
     _test_mask_from_regridder,
@@ -63,3 +64,9 @@ def test_mask_from_regridder(mask_keyword):
 def test_non_degree_crs():
     """Test for coordinates with non-degree units."""
     _test_non_degree_crs(ESMFBilinear)
+
+
+def test_dtype_handling(src_tgt_types, in_dtype):
+    """Test regridding scheme handles dtype as expected."""
+    src_type, tgt_type = src_tgt_types
+    _test_dtype_handling(ESMFBilinear, src_type, tgt_type, in_dtype)